@@ -2,15 +2,13 @@ name: camera
 description: A Flutter plugin for getting information about and controlling the
   camera on Android and iOS. Supports previewing the camera feed, capturing images, capturing video,
   and streaming image buffers to dart.
-version: 0.6.5
+version: 0.6.6
 homepage: https://github.com/flutter/plugins/tree/master/packages/camera/camera
 
 dependencies:
   flutter:
     sdk: flutter
-  #TODO(BeMacized): Replace with reference to pub.dev version once updated platform interface has been published.
-  camera_platform_interface:
-    path: ../camera_platform_interface
+  camera_platform_interface: ^1.5.0
   pedantic: ^1.8.0
 
 dev_dependencies:
@@ -33,5 +31,5 @@ flutter:
         pluginClass: CameraPlugin
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
