@@ -81,7 +81,7 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         {
           if (camera != null) {
             try {
-              camera.open();
+              camera.open(call.argument("imageFormatGroup"));
               result.success(null);
             } catch (Exception e) {
               handleException(e, result);
