@@ -15,4 +15,5 @@ export 'package:camera_platform_interface/camera_platform_interface.dart'
         ExposureMode,
         FocusMode,
         ResolutionPreset,
-        XFile;
+        XFile,
+        ImageFormatGroup;
