@@ -6,6 +6,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:camera_platform_interface/camera_platform_interface.dart';
 
 /// A single color plane of image data.
 ///
@@ -41,32 +42,6 @@ class Plane {
   final int width;
 }
 
-// TODO:(bmparr) Turn [ImageFormatGroup] to a class with int values.
-/// Group of image formats that are comparable across Android and iOS platforms.
-enum ImageFormatGroup {
-  /// The image format does not fit into any specific group.
-  unknown,
-
-  /// Multi-plane YUV 420 format.
-  ///
-  /// This format is a generic YCbCr format, capable of describing any 4:2:0
-  /// chroma-subsampled planar or semiplanar buffer (but not fully interleaved),
-  /// with 8 bits per color sample.
-  ///
-  /// On Android, this is `android.graphics.ImageFormat.YUV_420_888`. See
-  /// https://developer.android.com/reference/android/graphics/ImageFormat.html#YUV_420_888
-  ///
-  /// On iOS, this is `kCVPixelFormatType_420YpCbCr8BiPlanarVideoRange`. See
-  /// https://developer.apple.com/documentation/corevideo/1563591-pixel_format_identifiers/kcvpixelformattype_420ypcbcr8biplanarvideorange?language=objc
-  yuv420,
-
-  /// 32-bit BGRA.
-  ///
-  /// On iOS, this is `kCVPixelFormatType_32BGRA`. See
-  /// https://developer.apple.com/documentation/corevideo/1563591-pixel_format_identifiers/kcvpixelformattype_32bgra?language=objc
-  bgra8888,
-}
-
 /// Describes how pixels are represented in an image.
 class ImageFormat {
   ImageFormat._fromPlatformData(this.raw) : group = _asImageFormatGroup(raw);
@@ -86,9 +61,13 @@ class ImageFormat {
 
 ImageFormatGroup _asImageFormatGroup(dynamic rawFormat) {
   if (defaultTargetPlatform == TargetPlatform.android) {
-    // android.graphics.ImageFormat.YUV_420_888
-    if (rawFormat == 35) {
-      return ImageFormatGroup.yuv420;
+    switch (rawFormat) {
+      // android.graphics.ImageFormat.YUV_420_888
+      case 35:
+        return ImageFormatGroup.yuv420;
+      // android.graphics.ImageFormat.JPEG
+      case 256:
+        return ImageFormatGroup.jpeg;
     }
   }
 
